Split automations parsing.py into _yaml/_decompose/_ranges submodules
parsing.py was 896 lines, over the 800-line soft cap. Extract the two
cohesive concerns it mixed with the document walk:

- _decompose.py — handler-body → AutomationTree decomposition (the
  action/condition/render family plus _safe_tree/_block_tree).
- _ranges.py — ruamel lc → 1-indexed line ranges + raw_yaml slice dump.
- _yaml.py — the shared round-trip YAML factory both sides use.

parsing.py keeps the document walk that stitches them into the ordered
ParsedAutomation list (611 lines) and re-exports the moved names so
existing 'from .parsing import ...' callers (emitter, writing,
writing_lists) and the parser tests are unchanged. Behavior-free.
